requirements.txt
GitPython==2.1.11
python-slugify==2.0.1
six==1.12.0
